tests: ruff check --fix --unsafe-fixes (test-scoped config)
Applied ruff's full auto-fix set (safe + unsafe) against tests/ruff.toml:
3439 of 3810 findings fixed (RUF059 unused-unpack renames, F401 unused-import
removal, PERF401 list-comprehension rewrites, B007 loop-var renames, RUF005
concat rewrites, RUF015/RUF046/UP035/UP032/UP034 modernizations, etc.).
Verified: diff spot-checked for the historically-risky unsafe categories
(RUF012 mutable-class-default, B011 assert-False, RUF043 pytest.raises regex)
-- none fired in this diff. Full-repo pytest --collect-only succeeds (29519
tests, 0 collection errors); a 596-test sample across scorers/, grouped_cat_fe/,
and 2 standalone regression files passes clean.

371 findings remain (not auto-fixable); next pass triages those by hand.

+++ tests/feature_selection/_synthetic_distributions.py
@@ -46,7 +46,8 @@ still "lognormal-shaped" (or t-shaped, etc.), just relocated to a legal domain.
 
 from __future__ import annotations
 
-from typing import Callable, Dict, Iterable, Mapping, Sequence
+from typing import Callable, Dict
+from collections.abc import Iterable, Mapping, Sequence
 
 import numpy as np
 
@@ -182,7 +183,7 @@ def with_outliers(
     n = arr.size
     if n == 0 or frac <= 0.0:
         return arr
-    k = int(round(frac * n))
+    k = round(frac * n)
     k = max(1, min(k, n - 1))  # at least one outlier, at least one clean point
     q1, med, q3 = np.quantile(arr, [0.25, 0.5, 0.75])
     iqr = q3 - q1
